requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose
groq
requests
